linlin bugfix; moving some stuff around
